Src
this is a react app that try to copy a website. to use this simply install node js and then choose a folder then "create-react-app name"
then replace src with this src. this project not using react routing yet so to see other page simply put in the index.js put "import App2 from './Blog' " then replace <App /> to <App2 />.